feat(feature/task): The endpoints multipart format dates to Json was corrected

+++ src/main/java/com/elixirline/service/elixirline_backend/agriculturalactivitiesmanagement/tasks/domain/model/commands/CreateInputsCommand.java
@@ -6,7 +6,7 @@ import org.springframework.web.multipart.MultipartFile;
 public record CreateInputsCommand(
         String name,
         String description,
-        Long quantity,
+        Double quantity,
         Long winegrowerId,
         UnitType unit,
         MultipartFile imageFile
